# src/api/swagger.json
{
  "openapi": "3.0.0",
  "servers": [],
  "info": {
    "title": "",
    "description": "",
    "version": ""
  },
...
  "paths": {
...
    "/modules": {
      "get": {
        "tags": [
          "Module"
        ],
        "summary": "List all modules",
        "description": "Returns all modules",
        "operationId": "FindAllModules",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Module"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Module"
        ],
        "summary": "Create new module",
        "description": "",
        "operationId": "CreateModule",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Module"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/NewModule"
        }
      }
    },
    "/modules/{id}": {
      "get": {
        "tags": [
          "Module"
        ],
        "summary": "Get module",
        "description": "Returns the given module",
        "operationId": "FindModule",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of module to return",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Module"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Module"
        ],
        "summary": "Update module",
        "description": "Updates the given module",
        "operationId": "UpdateModule",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of module to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Module"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "$ref": "#/components/requestBodies/NewModule"
        }
      },
      "delete": {
        "tags": [
          "Module"
        ],
        "summary": "Delete module",
        "description": "Removes the given module",
        "operationId": "DeleteModule",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of module to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "requestBodies": {
...
      "NewModule": {
        "content": {
          "multipart/form-data": {
            "schema": {
              "$ref": "#/components/schemas/NewModule"
            }
          },
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/NewModule"
            }
          }
        },
        "description": "Module object that needs to be added to the database",
        "required": true
      },
...
    },
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "NewModule": {
        "type": "object",
        "properties": {
          "MName": {
            "type": "string"
          },
          "MShortName": {
            "type": "string"
          },
          "MType": {
            "type": "string"
          },
          "MTypeDescription": {
            "type": "string"
          },
          "MPLEPartNumber": {
            "type": "string"
          },
          "MRevision": {
            "type": "string"
          },
          "MDescription": {
            "type": "string"
          },
          "file": {
            "type": "string",
            "format": "binary"
          },
          "MNetSizeHeight": {
            "type": "integer"
          },
          "MNetSizeWidth": {
            "type": "integer"
          },
          "MNetSizeLength": {
            "type": "integer"
          },
          "MNetWeight": {
            "type": "integer"
          },
          "MGrossWeight": {
            "type": "integer"
          },
          "MMarginPercent": {
            "type": "integer"
          },
          "MEOL": {
            "type": "boolean"
          },
          "MTag": {
            "type": "string"
          },
          "components": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "services": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          },
          "assemblies": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      },
...
    }
  }
}